refactor: move upgrade code to separate modules
This results in clearer code.
